requirements.txt
Django==1.7.3
pyzmq==14.5.0
requests==2.5.1
